requirements.txt
openai
python-dotenv
chromadb
requests
beautifulsoup4
gradio